Move ExchangeRate to concordium-contracts-common
