+++ rust-src/concordium_base/src/base.rs
@@ -1,6 +1,8 @@
 //! Basis type definitions that are used throughout the crate.
 
-pub use concordium_contracts_common::{Address, ContractAddress, ContractIndex, ContractSubIndex};
+pub use concordium_contracts_common::{
+    Address, ContractAddress, ContractIndex, ContractSubIndex, ExchangeRate,
+};
 use crypto_common::{
     derive::{SerdeBase16Serialize, Serial, Serialize},
     deserial_string,
@@ -10,7 +12,6 @@ use crypto_common::{
 };
 use derive_more::{Add, Display, From, FromStr, Into};
 use id::types::VerifyKey;
-use num::Integer;
 use rand::{CryptoRng, Rng};
 use random_oracle::RandomOracle;
 use std::{
@@ -886,109 +887,6 @@ impl<T: Ord> InclusiveRange<T> {
     pub fn contains(&self, x: &T) -> bool { &self.min <= x && x <= &self.max }
 }
 
-/// An exchange rate between two quantities. This is never 0, and the exchange
-/// rate should also never be infinite.
-#[derive(PartialEq, Eq, SerdeSerialize, SerdeDeserialize, Serial, Debug, Clone, Copy)]
-#[serde(try_from = "exchange_rate_json::ExchangeRateJSON")]
-pub struct ExchangeRate {
-    #[serde(deserialize_with = "crate::internal::deserialize_non_default::deserialize")]
-    pub numerator:   u64,
-    #[serde(deserialize_with = "crate::internal::deserialize_non_default::deserialize")]
-    pub denominator: u64,
-}
-
-impl ExchangeRate {
-    /// Attempt to construct an exchange rate from a numerator and denominator.
-    /// The numerator and denominator must both be non-zero, and they have to be
-    /// in reduced form.
-    pub fn new(numerator: u64, denominator: u64) -> Option<Self> {
-        if numerator != 0 && denominator != 0 && num::integer::gcd(numerator, denominator) == 1 {
-            Some(Self {
-                numerator,
-                denominator,
-            })
-        } else {
-            None
-        }
-    }
-}
-
-#[derive(Debug, Error)]
-/// An error that may occur when converting from a string to an exchange rate.
-pub enum ExchangeRateConversionError {
-    #[error("Could not convert from decimal: {0}")]
-    FromDecimal(#[from] rust_decimal::Error),
-    #[error("Exchange rate must be strictly positive.")]
-    NotStrictlyPositive,
-    #[error("Exchange rate is not representable, either it is too large or has too many digits.")]
-    Unrepresentable,
-}
-
-impl FromStr for ExchangeRate {
-    type Err = ExchangeRateConversionError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut decimal = rust_decimal::Decimal::from_str_exact(s)?;
-        decimal.normalize_assign();
-        if decimal.is_zero() || decimal.is_sign_negative() {
-            return Err(ExchangeRateConversionError::NotStrictlyPositive);
-        }
-        let mantissa = decimal.mantissa();
-        let scale = decimal.scale();
-        let denominator = 10u64
-            .checked_pow(scale)
-            .ok_or(ExchangeRateConversionError::Unrepresentable)?;
-        let numerator: u64 = mantissa
-            .try_into()
-            .map_err(|_| ExchangeRateConversionError::Unrepresentable)?;
-        let g = numerator.gcd(&denominator);
-        Ok(ExchangeRate {
-            numerator:   numerator / g,
-            denominator: denominator / g,
-        })
-    }
-}
-
-mod exchange_rate_json {
-    use super::*;
-    #[derive(SerdeDeserialize)]
-    #[serde(untagged)]
-    pub enum ExchangeRateJSON {
-        String(String),
-        Num(f64),
-        Object { numerator: u64, denominator: u64 },
-    }
-
-    impl TryFrom<ExchangeRateJSON> for ExchangeRate {
-        type Error = ExchangeRateConversionError;
-
-        fn try_from(value: ExchangeRateJSON) -> Result<Self, Self::Error> {
-            match value {
-                ExchangeRateJSON::String(value) => value.parse(),
-                ExchangeRateJSON::Num(v) => v.to_string().parse(),
-                ExchangeRateJSON::Object {
-                    numerator,
-                    denominator,
-                } => {
-                    let g = numerator.gcd(&denominator);
-                    Ok(ExchangeRate {
-                        numerator:   numerator / g,
-                        denominator: denominator / g,
-                    })
-                }
-            }
-        }
-    }
-}
-
-impl Deserial for ExchangeRate {
-    fn deserial<R: ReadBytesExt>(source: &mut R) -> ParseResult<Self> {
-        let numerator = source.get()?;
-        let denominator = source.get()?;
-        Self::new(numerator, denominator).ok_or_else(|| anyhow::anyhow!("Invalid exchange rate."))
-    }
-}
-
 #[derive(SerdeSerialize, SerdeDeserialize, Serial, Debug, Clone, Copy)]
 #[serde(try_from = "leverage_factor_json::LeverageFactorRaw")]
 /// The amount of leverage that a baker can get from delegation. A leverage
@@ -1387,41 +1285,5 @@ mod tests {
             "Case 4."
         );
         assert!(serde_json::from_str::<PartsPerHundredThousands>("0.123456").is_err());
-    }
-
-    #[test]
-    fn test_exchange_rate_json() {
-        let data = ExchangeRate {
-            numerator:   1,
-            denominator: 100,
-        };
-        assert_eq!(
-            data,
-            serde_json::from_str("{\"numerator\": 1, \"denominator\": 100}").unwrap(),
-            "Exchange rate: case 1"
-        );
-        assert_eq!(
-            data,
-            serde_json::from_value(serde_json::from_str("0.01").unwrap()).unwrap(),
-            "Exchange rate: case 2"
-        );
-        let data2 = ExchangeRate {
-            numerator:   10,
-            denominator: 1,
-        };
-        assert_eq!(
-            data2,
-            serde_json::from_str("10").unwrap(),
-            "Exchange rate: case 3"
-        );
-        let data3 = ExchangeRate {
-            numerator:   17,
-            denominator: 39,
-        };
-        assert_eq!(
-            data3,
-            serde_json::from_str(&serde_json::to_string(&data3).unwrap()).unwrap(),
-            "Exchange rate: case 4"
-        );
     }
 }
